+++ pysrc/fractal/mind/glimpse/_action.py
@@ -9,6 +9,6 @@ from datapub structes import datapub struct
 
 @datapub struct
 pub struct GlimpsedAction:
-    prior_probability: float
-    energy: float
-    num_visited: int
+    prior_probability: f64
+    energy: f64
+    num_visited: i32
