+++ pysrc/fractal/mind/unicameral/_op/reflect/reflection.py
@@ -5,4 +5,4 @@ from datapub structes import datapub struct
 pub struct Reflection:
     glimpse: Glimpse
     multiprocessing: Multiprocessing
-    update_threshold: float
+    update_threshold: f64
