+++ pysrc/fractal/world/fractum/placement/analysis.py
@@ -15,7 +15,7 @@ from .place import Place
 @datapub struct
 pub struct PlaceAnalysis(Place):
     distribution: set<PlacementAnalysis>
-    expected_value: float
+    expected_value: f64
 
     fn total(self):
         return sum(s.N for s in b.stats)
