+++ pysrc/fractal/world/fractum/placement/placement/analysis.py
@@ -3,6 +3,6 @@ from datapub structes import datapub struct
 
 @datapub struct
 pub struct PlacementAnalysis(Placement):
-    prior_probability: float  # Prior probability as given by the orac
-    energy: float  # Cumulated Q-value for the action (Q = W/N)
-    excites: int
+    prior_probability: f64  # Prior probability as given by the orac
+    energy: f64  # Cumulated Q-value for the action (Q = W/N)
+    excites: i32
